Check R set up before a class

Checks that the required versions of R, RStudio, R packages and other dependencies are installed.

chk_requirements(path = system.file("default.yaml", package = "checker"))

Arguments

path

path to yaml file. Defaults to a file that comes with the package.

Details

chk_requirements() checks that the computer set up before class. It check

R version

RStudio version

RStudio options

R packages are installed (with version if necessary)

git version

quarto version

These requirements are specified in a yaml file specified by the path argument which can be on the users computer or at a URL. If not set, the function defaults to using a built-in yaml file, which may not require the latest version.

Examples

chk_requirements()
#> → Date = 2023-03-13 19:19:16
#> → os = Ubuntu 18.04.6 LTS
#> ✖ Are you using RStudio?
#> ✔ R version 4.2.2 is installed
#> ✔ quarto version 1.3.262 is installed
#> ✔ git version 2.17.1 is installed
#> ✔ tidyverse version 1.3.2 is installed
#> ✔ here version 1.0.1 is installed
#> ✔ quarto version 1.2 is installed
#> ✖ You have some issues that need addressing